feat: added charting functionality for historical data

+++ blazor-crypto-marketcap/Data/CoinMetrics.cs
@@ -5,12 +5,12 @@ namespace blazor_crypto_marketcap.Data
 {
     public class USD
     {
-        public double? price { get; set; }
-        public long? volume_24h { get; set; }
-        public double? percent_change_1h { get; set; }
-        public double? percent_change_24h { get; set; }
-        public double? percent_change_7d { get; set; }
-        public long? market_cap { get; set; }
+        public decimal? price { get; set; }
+        public decimal? volume_24h { get; set; }
+        public decimal? percent_change_1h { get; set; }
+        public decimal? percent_change_24h { get; set; }
+        public decimal? percent_change_7d { get; set; }
+        public decimal? market_cap { get; set; }
         public DateTime last_updated { get; set; }
     }
